TweetAgent/requirements.txt
langchain_core==0.3.30
langchain_openai==0.3.0
langgraph==0.2.64
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2
